Improved Island Generation
Added high smoothing so levels from -0.2 to 0.2 are smoother and flatter.

Improved Island shape aka (noise mask) by changing out a subtraction of values or multiplication with mask to a lerp between the noise between 1 to -1 and the noise between 0 and -1 mapped to an ln curve where 1->0 and -1->-1.

Added change see button