Continued Ocean Tile creation progress
Player height 1.68->1.7 to make conversion to foot elevation easier.
Ocean shader shows even odd pattern for tiles for debuging.
Added calculation of which corner of frustum bound ocean plane is located closest to player and the location of each other corner in relation to that one.
Added calculation for center of said plane for finding out where the inside vs outside of the plane is.